Support case-insensitive filesystems
Allow people working on case-insensitive filesytem edit files in the
content directory. The GitHub action can be used to build the github
pages, which still creates case-sensitive directory names such as
field/deliverable and field/Deliverable.

The filename of the content/field/deliverable.md isn't significant
as the fieldname is take from the frontmatter field value.