Removed inappropriate semantic versioning
Removed the version numbering scheme which
does not match the HTML5/WHATWG approach
we take for standards development.

Ironically, the mecahnism for building older
specification documents was failing as the specification
data was still under development.

Our specifications are living documents,
older versions are visible in GitHub
and on the National Archives.

The specifications when are published to GOV.UK where
they are change-controlled using the publication date,
which is what should be cited when refering to a specific
version of the specification document.

Changes to the specification are explained in the specification.
A field in a published specification will be visible by having
a replacement field, and being explained in the text of the
specification.